refactor(web): extract console/spa-mount/ws-auth mixins from web_server.py (shard s5 union)
