+++ hermes_cli/web_server.py
@@ -134,6 +134,35 @@ except ImportError:
 
 WEB_DIST = Path(os.environ["HERMES_WEB_DIST"]) if "HERMES_WEB_DIST" in os.environ else Path(__file__).parent / "web_dist"
 _log = logging.getLogger(__name__)
+from hermes_cli.ws_auth_gate_mixin import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
+    _LOOPBACK_HOSTS,
+    _ws_auth_mode,
+    _ws_auth_ok,
+    _ws_auth_reason,
+    _ws_client_is_allowed,
+    _ws_client_reason,
+    _ws_close_reason,
+    _ws_host_origin_is_allowed,
+    _ws_host_origin_reason,
+    _ws_request_is_allowed,
+    _ws_request_reason,
+)
+from hermes_cli.spa_mount_mixin import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
+    _IMMUTABLE_ASSET_CACHE_CONTROL,
+    _normalise_prefix,
+    _render_active_theme_bootstrap_css,
+    mount_spa,
+)
+
+from hermes_cli.console_ws_mixin import (  # noqa: E402
+    _console_json_payload,
+    _console_profile_from_ws,
+    _console_send,
+    _console_send_result,
+    _execute_console_line,
+    _get_console_executor,
+    console_ws,
+)
 
 # ---------------------------------------------------------------------------
 # Per-channel subscriber registry used by /api/pub (PTY-side gateway → dashboard)
@@ -8412,7 +8441,6 @@ from hermes_cli.web_routers.whatsapp_onboarding import (  # noqa: E402,F401 — 
     _whatsapp_session_path,
     _spawn_whatsapp_pairing_process,
     _watch_whatsapp_pairing,
-    _write_platform_enabled,
     apply_whatsapp_onboarding,
     cancel_whatsapp_onboarding,
     get_whatsapp_onboarding_status,
@@ -14025,234 +14053,7 @@ async def _legacy_pump(ws: "WebSocket", bridge) -> None:
 
 
 _VALID_CHANNEL_RE = re.compile(r"^[A-Za-z0-9._-]{1,128}$")
-# Starlette's TestClient reports the peer as "testclient"; treat it as
-# loopback so tests don't need to rewrite request scope.
-_LOOPBACK_HOSTS = frozenset({"127.0.0.1", "::1", "localhost", "testclient"})
 
-
-def _ws_client_reason(ws: "WebSocket") -> Optional[str]:
-    """Return a rejection reason for the client IP, or None when allowed.
-
-    Reasons are short machine-parseable tokens logged on the rejection path
-    so a "WS keeps closing" report can be diagnosed from agent.log without a
-    repro. ``None`` means the peer IP passed this gate.
-
-    See :func:`_ws_client_is_allowed` for the full policy rationale.
-    """
-    if getattr(app.state, "auth_required", False):
-        return None
-    bound_host = (getattr(app.state, "bound_host", "") or "").strip().lower()
-    if bound_host and bound_host not in _LOOPBACK_HOSTS:
-        return None
-    client_host = ws.client.host if ws.client else ""
-    if not client_host:
-        # Fail-closed: a loopback-bound dashboard with auth disabled must
-        # not accept a WebSocket with no identifiable peer. ASGI servers
-        # behind a misconfigured proxy or unix socket can deliver
-        # ws.client == None or "" — treating that as "allowed" would let
-        # an unidentified peer reach a loopback-only surface.
-        return f"missing_or_empty_peer bound={bound_host or '?'}"
-    if client_host in _LOOPBACK_HOSTS:
-        return None
-    return f"peer_not_loopback peer={client_host} bound={bound_host or '?'}"
-
-
-def _ws_client_is_allowed(ws: "WebSocket") -> bool:
-    """Check if the WebSocket client IP is acceptable.
-
-    Loopback bind: only loopback clients allowed — the legacy
-    ``?token=<_SESSION_TOKEN>`` path is the only auth we have, so we
-    don't want LAN hosts guessing tokens.
-
-    Explicit non-loopback bind (``--host 0.0.0.0``, ``--host ::``, or a
-    specific address such as a Tailscale/LAN IP, always with
-    ``--insecure``): allow any peer. The operator explicitly opted into
-    non-loopback exposure, so the loopback-only peer restriction does not
-    apply. DNS-rebinding is still blocked by the Host/Origin guard in
-    :func:`_ws_host_origin_is_allowed`, which mirrors the HTTP layer and
-    requires the Host header to match the bound interface — the same
-    defence ``_is_accepted_host`` applies to non-loopback HTTP requests.
-
-    Gated mode: any peer is allowed — uvicorn's ``proxy_headers=True``
-    (enabled when the OAuth gate is active so cookies can pick up
-    ``X-Forwarded-Proto``) rewrites ``ws.client.host`` to the
-    X-Forwarded-For value, which is the real internet client IP. The
-    OAuth gate + single-use ``?ticket=`` is the auth at that point; the
-    Host/Origin guard in :func:`_ws_host_origin_is_allowed` is what
-    blocks DNS-rebinding here, not the peer IP.
-    """
-    if getattr(app.state, "auth_required", False):
-        return True
-    # Any explicit non-loopback bind (0.0.0.0, ::, or a specific LAN /
-    # Tailscale address) means the operator opted into non-loopback
-    # access via --insecure.  The loopback-only peer gate only applies to
-    # an actual loopback bind; otherwise the WS handshake is rejected even
-    # though same-bind HTTP requests pass _is_accepted_host.
-    bound_host = (getattr(app.state, "bound_host", "") or "").strip().lower()
-    if bound_host and bound_host not in _LOOPBACK_HOSTS:
-        return True
-    client_host = ws.client.host if ws.client else ""
-    if not client_host:
-        # Fail-closed: see _ws_client_reason for rationale. An empty
-        # client_host on a loopback-bound dashboard with auth disabled
-        # must be rejected, not accepted as a default-allow.
-        return False
-    return client_host in _LOOPBACK_HOSTS
-
-
-def _ws_host_origin_reason(ws: "WebSocket") -> Optional[str]:
-    """Return a Host/Origin rejection reason, or None when allowed.
-
-    Mirrors :func:`_ws_host_origin_is_allowed` but yields a short
-    machine-parseable token (``host_mismatch …`` / ``origin_mismatch …``)
-    on rejection so the close path can log *why* the upgrade was refused.
-    """
-    bound_host = getattr(app.state, "bound_host", None)
-    if not bound_host:
-        return None
-
-    host_header = ws.headers.get("host", "")
-    if not _is_accepted_host(host_header, bound_host):
-        return f"host_mismatch host={host_header or '?'} bound={bound_host}"
-
-    origin = ws.headers.get("origin", "")
-    if not origin:
-        return None
-
-    parsed = urllib.parse.urlparse(origin)
-    if parsed.scheme not in {"http", "https"}:
-        # Non-web origin (packaged Electron: file://, null, app://). The
-        # upstream credential check is the real auth boundary; trust it.
-        # See _ws_host_origin_is_allowed for the full rationale.
-        return None
-
-    if not parsed.netloc:
-        return f"origin_mismatch origin={origin} bound={bound_host}"
-
-    if not _is_accepted_host(parsed.netloc, bound_host):
-        return f"origin_mismatch origin={origin} bound={bound_host}"
-    return None
-
-
-def _ws_host_origin_is_allowed(ws: "WebSocket") -> bool:
-    """Apply the dashboard Host/Origin guard to WebSocket upgrades.
-
-    FastAPI HTTP middleware does not run for WebSocket routes, so the
-    DNS-rebinding Host check used for normal dashboard HTTP requests must be
-    repeated here before accepting the upgrade.  Browsers also send an Origin
-    header on WebSocket handshakes; when present, require it to target the
-    same bound dashboard host.
-    """
-    return _ws_host_origin_reason(ws) is None
-
-
-def _ws_request_reason(ws: "WebSocket") -> Optional[str]:
-    """First Host/Origin or peer-IP rejection reason, or None when allowed."""
-    return _ws_host_origin_reason(ws) or _ws_client_reason(ws)
-
-
-def _ws_request_is_allowed(ws: "WebSocket") -> bool:
-    """Return True when the WebSocket upgrade matches dashboard boundaries."""
-    return _ws_host_origin_is_allowed(ws) and _ws_client_is_allowed(ws)
-
-
-def _ws_auth_mode() -> str:
-    """Short label for the active WS auth mode — logged on every connection."""
-    if getattr(app.state, "auth_required", False):
-        return "gated"
-    bound_host = (getattr(app.state, "bound_host", "") or "").strip().lower()
-    if bound_host and bound_host not in _LOOPBACK_HOSTS:
-        return "insecure"
-    return "loopback"
-
-
-def _ws_auth_reason(ws: "WebSocket") -> tuple[Optional[str], str]:
-    """Validate WS-upgrade auth; return ``(reason, credential)``.
-
-    ``reason`` is None when the credential is accepted, else a short
-    machine-parseable token explaining the rejection (``no_credential``,
-    ``token_mismatch``, ``ticket_invalid``, ``internal_invalid``).
-    ``credential`` names which credential type was presented (``ticket``,
-    ``internal``, ``token``, or ``none``) so the accepted path can log *how*
-    a peer authed, not just that it did.
-
-    Loopback / ``--insecure``: legacy ``?token=<_SESSION_TOKEN>`` query
-    parameter, constant-time compared.
-
-    Gated (public bind, no ``--insecure``): one of two credentials —
-
-    * ``?ticket=<single-use>`` — a browser-minted, single-use, 30s-TTL ticket
-      consumed against the dashboard-auth ticket store. This is what the SPA
-      (and native clients) use.
-    * ``?internal=<process-credential>`` — the process-lifetime internal
-      credential, used only by WS clients the server spawns itself (the
-      embedded-TUI PTY child attaching to ``/api/ws`` and ``/api/pub``). It
-      is multi-use and never expires so the child can reconnect, and is never
-      injected into the SPA — see ``dashboard_auth.ws_tickets`` for the
-      threat model.
-
-    The legacy ``?token=`` path is unconditionally rejected in gated mode
-    (the SPA bundle isn't carrying the token any longer, and a leaked
-    ``_SESSION_TOKEN`` must not grant WS access once the gate is engaged).
-
-    Audit-logs the rejection so operators can debug "WS keeps closing"
-    issues from the log.
-    """
-    auth_required = bool(getattr(app.state, "auth_required", False))
-    if auth_required:
-        # Lazy import — keeps this function importable in test harnesses
-        # that don't bring in the dashboard_auth layer.
-        from hermes_cli.dashboard_auth.audit import AuditEvent, audit_log
-        from hermes_cli.dashboard_auth.ws_tickets import (
-            TicketInvalid,
-            consume_internal_credential,
-            consume_ticket,
-        )
-
-        # Server-spawned children (PTY child → /api/ws, /api/pub) present the
-        # multi-use internal credential rather than a single-use ticket, so
-        # they survive reconnects and slow cold boots.
-        internal = ws.query_params.get("internal", "")
-        if internal:
-            try:
-                consume_internal_credential(internal)
-                return None, "internal"
-            except TicketInvalid as exc:
-                audit_log(
-                    AuditEvent.WS_TICKET_REJECTED,
-                    reason=f"internal: {exc}",
-                    ip=(ws.client.host if ws.client else ""),
-                    path=ws.url.path,
-                )
-                return "internal_invalid", "internal"
-
-        ticket = ws.query_params.get("ticket", "")
-        if not ticket:
-            return "no_credential", "none"
-
-        try:
-            consume_ticket(ticket)
-            return None, "ticket"
-        except TicketInvalid as exc:
-            audit_log(
-                AuditEvent.WS_TICKET_REJECTED,
-                reason=str(exc),
-                ip=(ws.client.host if ws.client else ""),
-                path=ws.url.path,
-            )
-            return "ticket_invalid", "ticket"
-
-    token = ws.query_params.get("token", "")
-    if not token:
-        return "no_credential", "none"
-    if hmac.compare_digest(token.encode(), _SESSION_TOKEN.encode()):
-        return None, "token"
-    return "token_mismatch", "token"
-
-
-def _ws_auth_ok(ws: "WebSocket") -> bool:
-    """True when the WS-upgrade credential is accepted. See _ws_auth_reason."""
-    return _ws_auth_reason(ws)[0] is None
 
 # Per-channel subscriber registry used by /api/pub (PTY-side gateway → dashboard)
 # and /api/events (dashboard → browser sidebar).  Keyed by an opaque channel id
@@ -14568,582 +14369,12 @@ def _forget_active_session_file(path: Path) -> None:
         pass
 
 
-def _ws_close_reason(text: str) -> str:
-    """Clamp a WS close reason to the protocol's 123-byte UTF-8 limit.
-
-    RFC 6455 caps the close-frame reason at 123 bytes; uvicorn raises if a
-    longer string is passed. Our reasons embed an attacker-controlled origin,
-    so truncate defensively rather than crash the close handler.
-    """
-    encoded = text.encode("utf-8", "replace")
-    if len(encoded) <= 123:
-        return text
-    return encoded[:120].decode("utf-8", "ignore") + "..."
-
-
 # ---------------------------------------------------------------------------
 # /api/console — safe Hermes Console command WebSocket.
 #
 # Unlike /api/pty, this endpoint never spawns a PTY, shell, or full Hermes CLI
 # subprocess. It runs the curated console engine in-process and exchanges
 # structured JSON frames with the dashboard xterm overlay.
-# ---------------------------------------------------------------------------
-
-_CONSOLE_PROMPT = "hermes> "
-_CONSOLE_COMMAND_TIMEOUT_SECONDS = 60.0
-_CONSOLE_OUTPUT_LIMIT = 50000
-
-# Console commands run in a worker thread. On a timeout, asyncio.wait_for cancels
-# the *awaitable*, but Python threads aren't preemptible, so a genuinely stuck
-# worker keeps running to completion. To keep that from exhausting the shared
-# default thread pool (asyncio.to_thread), we run console commands on a small
-# dedicated, bounded pool: a leaked worker is capped, and concurrent console
-# execution is bounded to a fixed number of threads regardless of reconnects.
-_CONSOLE_EXECUTOR_MAX_WORKERS = 4
-_console_executor: Optional[concurrent.futures.ThreadPoolExecutor] = None
-_console_executor_lock = threading.Lock()
-
-
-def _get_console_executor() -> concurrent.futures.ThreadPoolExecutor:
-    """Lazily create the bounded console worker pool (once per process)."""
-    global _console_executor
-    if _console_executor is None:
-        with _console_executor_lock:
-            if _console_executor is None:
-                _console_executor = concurrent.futures.ThreadPoolExecutor(
-                    max_workers=_CONSOLE_EXECUTOR_MAX_WORKERS,
-                    thread_name_prefix="hermes-console",
-                )
-                # Ensure the pool is torn down on interpreter exit. Don't wait on
-                # in-flight workers: a stuck 60s console command must not block
-                # shutdown (cancel_futures drops anything not yet started).
-                atexit.register(
-                    lambda: _console_executor
-                    and _console_executor.shutdown(wait=False, cancel_futures=True)
-                )
-    return _console_executor
-
-
-def _console_profile_from_ws(ws: WebSocket) -> Optional[str]:
-    profile = (ws.query_params.get("profile") or "").strip()
-    return profile or None
-
-
-def _execute_console_line(
-    engine: Any,
-    line: str,
-    *,
-    confirmed: bool,
-    profile: Optional[str],
-) -> Any:
-    # _profile_scope swaps process-global skill module paths; keep it inside
-    # the worker thread and never hold it across awaits.
-    with _profile_scope(profile):
-        return engine.execute(line, confirmed=confirmed)
-
-
-async def _console_send(
-    ws: WebSocket,
-    send_lock: asyncio.Lock,
-    payload: Dict[str, Any],
-) -> None:
-    async with send_lock:
-        await ws.send_json(payload)
-
-
-async def _console_send_result(
-    ws: WebSocket,
-    send_lock: asyncio.Lock,
-    result: Any,
-    *,
-    command_id: int,
-) -> None:
-    command = result.command or ""
-    status = result.status
-    if status == "ok":
-        if result.output:
-            await _console_send(
-                ws,
-                send_lock,
-                {
-                    "type": "output",
-                    "id": command_id,
-                    "stream": "stdout",
-                    "data": result.output,
-                    "command": command,
-                },
-            )
-        await _console_send(
-            ws,
-            send_lock,
-            {
-                "type": "complete",
-                "id": command_id,
-                "status": "ok",
-                "command": command,
-                "prompt": _CONSOLE_PROMPT,
-            },
-        )
-        return
-
-    if status == "error":
-        await _console_send(
-            ws,
-            send_lock,
-            {
-                "type": "error",
-                "id": command_id,
-                "message": result.output or "Command failed.",
-                "command": command,
-            },
-        )
-        await _console_send(
-            ws,
-            send_lock,
-            {
-                "type": "complete",
-                "id": command_id,
-                "status": "error",
-                "command": command,
-                "prompt": _CONSOLE_PROMPT,
-            },
-        )
-        return
-
-    if status == "confirm_required":
-        await _console_send(
-            ws,
-            send_lock,
-            {
-                "type": "confirm_required",
-                "id": command_id,
-                "command": command,
-                "message": result.confirmation_message or f"Run `{command}`?",
-                "prompt": _CONSOLE_PROMPT,
-            },
-        )
-        await _console_send(
-            ws,
-            send_lock,
-            {
-                "type": "complete",
-                "id": command_id,
-                "status": "confirm_required",
-                "command": command,
-                "prompt": _CONSOLE_PROMPT,
-            },
-        )
-        return
-
-    if status == "clear":
-        await _console_send(ws, send_lock, {"type": "clear", "id": command_id})
-        await _console_send(
-            ws,
-            send_lock,
-            {
-                "type": "complete",
-                "id": command_id,
-                "status": "clear",
-                "command": command,
-                "prompt": _CONSOLE_PROMPT,
-            },
-        )
-        return
-
-    if status == "exit":
-        await _console_send(
-            ws,
-            send_lock,
-            {
-                "type": "complete",
-                "id": command_id,
-                "status": "exit",
-                "command": command,
-                "prompt": "",
-            },
-        )
-        return
-
-    await _console_send(
-        ws,
-        send_lock,
-        {
-            "type": "error",
-            "id": command_id,
-            "message": f"Unknown console result status: {status}",
-            "command": command,
-        },
-    )
-
-
-def _console_json_payload(msg: Any) -> tuple[Optional[dict[str, Any]], Optional[str]]:
-    raw: str | bytes | None = msg.get("text")
-    if raw is None:
-        raw = msg.get("bytes")
-    if raw is None:
-        return None, None
-    if isinstance(raw, bytes):
-        try:
-            raw = raw.decode("utf-8")
-        except UnicodeDecodeError:
-            return None, "Console frames must be UTF-8 JSON."
-    try:
-        payload = json.loads(raw)
-    except json.JSONDecodeError:
-        return None, "Console frames must be JSON objects."
-    if not isinstance(payload, dict):
-        return None, "Console frames must be JSON objects."
-    return payload, None
-
-
-@app.websocket("/api/console")
-async def console_ws(ws: WebSocket) -> None:
-    peer = ws.client.host if ws.client else "?"
-
-    if not _DASHBOARD_EMBEDDED_CHAT_ENABLED:
-        _log.info("console refused: embedded chat disabled peer=%s", peer)
-        await ws.close(code=4404, reason="embedded chat disabled")
-        return
-
-    auth_reason, cred = _ws_auth_reason(ws)
-    mode = _ws_auth_mode()
-    if auth_reason is not None:
-        _log.warning(
-            "console auth rejected reason=%s mode=%s cred=%s peer=%s",
-            auth_reason, mode, cred, peer,
-        )
-        await ws.close(code=4401, reason=_ws_close_reason(f"auth: {auth_reason}"))
-        return
-
-    host_origin_reason = _ws_host_origin_reason(ws)
-    if host_origin_reason is not None:
-        _log.warning("console refused: %s peer=%s", host_origin_reason, peer)
-        await ws.close(code=4403, reason=_ws_close_reason(host_origin_reason))
-        return
-
-    client_reason = _ws_client_reason(ws)
-    if client_reason is not None:
-        _log.warning("console refused: %s", client_reason)
-        await ws.close(code=4408, reason=_ws_close_reason(client_reason))
-        return
-
-    await ws.accept()
-
-    profile = _console_profile_from_ws(ws)
-    send_lock = asyncio.Lock()
-
-    try:
-        from hermes_cli.console_engine import HermesConsoleEngine
-
-        engine = HermesConsoleEngine(output_limit=_CONSOLE_OUTPUT_LIMIT)
-        if profile and profile.lower() != "current":
-            _resolve_profile_dir(profile)
-    except HTTPException as exc:
-        await _console_send(
-            ws,
-            send_lock,
-            {
-                "type": "error",
-                "message": str(exc.detail),
-                "prompt": "",
-            },
-        )
-        await ws.close(code=4400, reason=_ws_close_reason(str(exc.detail)))
-        return
-    except Exception as exc:
-        _log.exception("console failed to initialize")
-        await _console_send(
-            ws,
-            send_lock,
-            {
-                "type": "error",
-                "message": f"Console unavailable: {exc}",
-                "prompt": "",
-            },
-        )
-        await ws.close(code=1011)
-        return
-
-    _log.info(
-        "console accepted peer=%s mode=%s cred=%s profile=%s",
-        peer,
-        mode,
-        cred,
-        profile or "current",
-    )
-    await _console_send(
-        ws,
-        send_lock,
-        {
-            "type": "ready",
-            "profile": profile or "current",
-            "prompt": _CONSOLE_PROMPT,
-        },
-    )
-
-    active_task: asyncio.Task | None = None
-    pending_confirmation: Optional[str] = None
-    command_generation = 0
-
-    async def run_command(line: str, *, confirmed: bool, command_id: int) -> None:
-        nonlocal active_task, pending_confirmation, command_generation
-        try:
-            loop = asyncio.get_running_loop()
-            result = await asyncio.wait_for(
-                loop.run_in_executor(
-                    _get_console_executor(),
-                    functools.partial(
-                        _execute_console_line,
-                        engine,
-                        line,
-                        confirmed=confirmed,
-                        profile=profile,
-                    ),
-                ),
-                timeout=_CONSOLE_COMMAND_TIMEOUT_SECONDS,
-            )
-        except asyncio.CancelledError:
-            raise
-        except asyncio.TimeoutError:
-            if command_id == command_generation:
-                pending_confirmation = None
-                await _console_send(
-                    ws,
-                    send_lock,
-                    {
-                        "type": "error",
-                        "id": command_id,
-                        "message": (
-                            "Command timed out. Hermes Console returned to the prompt."
-                        ),
-                        "command": line,
-                    },
-                )
-                await _console_send(
-                    ws,
-                    send_lock,
-                    {
-                        "type": "complete",
-                        "id": command_id,
-                        "status": "timeout",
-                        "command": line,
-                        "prompt": _CONSOLE_PROMPT,
-                    },
-                )
-        except Exception as exc:
-            if command_id == command_generation:
-                pending_confirmation = None
-                _log.exception("console command failed")
-                await _console_send(
-                    ws,
-                    send_lock,
-                    {
-                        "type": "error",
-                        "id": command_id,
-                        "message": str(exc) or exc.__class__.__name__,
-                        "command": line,
-                    },
-                )
-                await _console_send(
-                    ws,
-                    send_lock,
-                    {
-                        "type": "complete",
-                        "id": command_id,
-                        "status": "error",
-                        "command": line,
-                        "prompt": _CONSOLE_PROMPT,
-                    },
-                )
-        else:
-            if command_id != command_generation:
-                return
-            pending_confirmation = (
-                result.command if result.status == "confirm_required" else None
-            )
-            await _console_send_result(
-                ws,
-                send_lock,
-                result,
-                command_id=command_id,
-            )
-            if result.status == "exit":
-                await ws.close(code=1000)
-        finally:
-            if command_id == command_generation:
-                active_task = None
-
-    async def start_command(line: str, *, confirmed: bool = False) -> None:
-        nonlocal active_task, command_generation
-        command_generation += 1
-        command_id = command_generation
-        active_task = asyncio.create_task(
-            run_command(line, confirmed=confirmed, command_id=command_id)
-        )
-
-    try:
-        while True:
-            try:
-                msg = await ws.receive()
-            except RuntimeError:
-                break
-            msg_type = msg.get("type")
-            if msg_type == "websocket.disconnect":
-                break
-
-            payload, error = _console_json_payload(msg)
-            if error:
-                await _console_send(
-                    ws,
-                    send_lock,
-                    {
-                        "type": "error",
-                        "message": error,
-                        "prompt": _CONSOLE_PROMPT,
-                    },
-                )
-                continue
-            if payload is None:
-                continue
-
-            frame_type = str(payload.get("type") or "").strip().lower()
-            if frame_type == "ping":
-                await _console_send(
-                    ws,
-                    send_lock,
-                    {
-                        "type": "pong",
-                        "prompt": _CONSOLE_PROMPT,
-                    },
-                )
-                continue
-
-            if frame_type == "cancel":
-                if active_task and not active_task.done():
-                    command_generation += 1
-                    active_task.cancel()
-                    active_task = None
-                    pending_confirmation = None
-                    await _console_send(
-                        ws,
-                        send_lock,
-                        {
-                            "type": "complete",
-                            "status": "cancelled",
-                            "prompt": _CONSOLE_PROMPT,
-                        },
-                    )
-                elif pending_confirmation:
-                    pending_confirmation = None
-                    await _console_send(
-                        ws,
-                        send_lock,
-                        {
-                            "type": "complete",
-                            "status": "cancelled",
-                            "prompt": _CONSOLE_PROMPT,
-                        },
-                    )
-                else:
-                    await _console_send(
-                        ws,
-                        send_lock,
-                        {
-                            "type": "complete",
-                            "status": "idle",
-                            "prompt": _CONSOLE_PROMPT,
-                        },
-                    )
-                continue
-
-            if active_task and not active_task.done():
-                await _console_send(
-                    ws,
-                    send_lock,
-                    {
-                        "type": "error",
-                        "message": "A console command is already running.",
-                        "prompt": _CONSOLE_PROMPT,
-                    },
-                )
-                continue
-
-            if frame_type == "confirm":
-                command = str(payload.get("command") or pending_confirmation or "").strip()
-                if not pending_confirmation:
-                    await _console_send(
-                        ws,
-                        send_lock,
-                        {
-                            "type": "error",
-                            "message": "No command is waiting for confirmation.",
-                            "prompt": _CONSOLE_PROMPT,
-                        },
-                    )
-                    continue
-                if command != pending_confirmation:
-                    await _console_send(
-                        ws,
-                        send_lock,
-                        {
-                            "type": "error",
-                            "message": "Confirmation does not match the pending command.",
-                            "prompt": _CONSOLE_PROMPT,
-                        },
-                    )
-                    continue
-                pending_confirmation = None
-                await start_command(command, confirmed=True)
-                continue
-
-            if frame_type in {"input", "command"}:
-                line = str(payload.get("line") or payload.get("command") or "").strip()
-                if not line:
-                    await _console_send(
-                        ws,
-                        send_lock,
-                        {
-                            "type": "complete",
-                            "status": "ok",
-                            "prompt": _CONSOLE_PROMPT,
-                        },
-                    )
-                    continue
-                if pending_confirmation:
-                    await _console_send(
-                        ws,
-                        send_lock,
-                        {
-                            "type": "error",
-                            "message": (
-                                "Confirm or cancel the pending command before "
-                                "running another one."
-                            ),
-                            "prompt": _CONSOLE_PROMPT,
-                        },
-                    )
-                    continue
-                await start_command(line)
-                continue
-
-            await _console_send(
-                ws,
-                send_lock,
-                {
-                    "type": "error",
-                    "message": f"Unsupported console frame: {frame_type or '?'}",
-                    "prompt": _CONSOLE_PROMPT,
-                },
-            )
-    except WebSocketDisconnect:
-        pass
-    finally:
-        if active_task and not active_task.done():
-            active_task.cancel()
-            try:
-                await active_task
-            except (asyncio.CancelledError, Exception):
-                pass
 
 
 @app.websocket("/api/pty")
@@ -15434,272 +14665,6 @@ async def events_ws(ws: WebSocket) -> None:
 
                 if not subs:
                     event_channels.pop(channel, None)
-
-
-def _normalise_prefix(raw: Optional[str]) -> str:
-    """Normalise an X-Forwarded-Prefix header value.
-
-    Thin re-export of :func:`hermes_cli.dashboard_auth.prefix.normalise_prefix`
-    — the single source of truth lives in the dashboard_auth package so
-    the gate middleware, the OAuth routes, the cookie helpers, and the
-    SPA mount all agree on validation rules.
-    """
-    from hermes_cli.dashboard_auth.prefix import normalise_prefix
-    return normalise_prefix(raw)
-
-
-def _render_active_theme_bootstrap_css() -> str:
-    """Critical-CSS shim for the active user theme.
-
-    Returns a ``<style>`` block with the ``:root`` CSS variables that
-    ``ThemeProvider.applyTheme()`` installs once the
-    ``/api/dashboard/themes`` round-trip completes.  The goal is to
-    eliminate the green flash where the first paint shows the bundle's
-    default Hermes Teal canvas before the SPA flips the configured user
-    theme into place.
-
-    Built-in themes return an empty string — their full definitions live
-    in ``web/src/themes/presets.ts`` and are applied by the bundle
-    before paint, so no shim is needed for them.
-    """
-    try:
-        config = load_config()
-        active = cfg_get(config, "dashboard", "theme", default="default")
-        if not active or not isinstance(active, str):
-            return ""
-        # Built-in: the bundle already owns the definition, no flash.
-        if any(b["name"] == active for b in _BUILTIN_DASHBOARD_THEMES):
-            return ""
-        for theme in _discover_user_themes():
-            if theme.get("name") != active:
-                continue
-            palette = theme.get("palette") or {}
-            bg = palette.get("background") or {}
-            mg = palette.get("midground") or {}
-            bg_hex = bg.get("hex", "#0a0a0a") if isinstance(bg, dict) else "#0a0a0a"
-            mg_hex = mg.get("hex", "#e5e5e5") if isinstance(mg, dict) else "#e5e5e5"
-            typo = theme.get("typography") or {}
-            font_sans = typo.get("fontSans") or _THEME_DEFAULT_TYPOGRAPHY["fontSans"]
-            base_size = typo.get("baseSize") or _THEME_DEFAULT_TYPOGRAPHY["baseSize"]
-            # Defensive ``</style>`` escape — current values are well-known
-            # hex/font strings, but this keeps the helper safe if it is
-            # later extended to ship user-authored CSS literals.
-            def _esc(s: str) -> str:
-                return str(s).replace("</", "<\\/")
-            # Variable names MUST match what the bundle actually consumes:
-            #   - ``--background-base`` / ``--midground-base`` come from
-            #     ``layerVars()`` in ``web/src/themes/context.tsx``.
-            #   - ``--theme-font-sans`` / ``--theme-base-size`` come from
-            #     ``typographyVars()`` there, and ``index.css`` applies them
-            #     via ``html{font-family:var(--theme-font-sans);
-            #     font-size:var(--theme-base-size)}``.
-            # The ``html,body`` canvas rule references the SAME variables
-            # instead of literal values so runtime theme switches stay
-            # live: ``applyTheme()`` writes these vars as inline styles on
-            # ``documentElement``, which outrank this stylesheet block in
-            # the cascade — the rule below re-resolves automatically and
-            # never goes stale when the user picks a different theme.
-            return (
-                '<style id="hermes-theme-bootstrap">'
-                ":root{"
-                f"--background-base:{_esc(bg_hex)};"
-                f"--midground-base:{_esc(mg_hex)};"
-                f"--theme-font-sans:{_esc(font_sans)};"
-                f"--theme-base-size:{_esc(base_size)};"
-                "}"
-                "html,body{background-color:var(--background-base);"
-                "color:var(--midground-base);"
-                "font-family:var(--theme-font-sans);"
-                "font-size:var(--theme-base-size);}"
-                "</style>"
-            )
-        return ""
-    except Exception:
-        _log.debug("theme bootstrap render failed", exc_info=True)
-        return ""
-
-
-# Hashed bundle assets (``/assets/<name>-<contenthash>.<ext>``) are immutable
-# by construction: any content change produces a new filename, and the entry
-# point (index.html) is served ``no-store`` so it always references the
-# current hashes. A year-long immutable cache lets browsers skip even the
-# revalidation round-trip on every dashboard load.
-_IMMUTABLE_ASSET_CACHE_CONTROL = "public, max-age=31536000, immutable"
-
-
-def mount_spa(application: FastAPI):
-    """Mount the built SPA. Falls back to index.html for client-side routing.
-
-    The session token is injected into index.html via a ``<script>`` tag so
-    the SPA can authenticate against protected API endpoints without a
-    separate (unauthenticated) token-dispensing endpoint.
-
-    When served behind a path-prefix reverse proxy (e.g.
-    ``mission-control.tilos.com/hermes/*`` -> local Caddy -> :9119), the
-    proxy injects ``X-Forwarded-Prefix: /hermes`` on every request. We
-    rewrite the served ``index.html`` so absolute asset URLs (``/assets/...``)
-    and the SPA's runtime ``__HERMES_BASE_PATH__`` honour that prefix
-    without rebuilding the bundle.
-    """
-    # `hermes serve` is the headless backend: it must NEVER serve the browser
-    # SPA, even if a dist is lying around from a prior `dashboard`/build. Take
-    # the no-frontend path so only the JSON-RPC/WS/API surface is reachable.
-    _headless = os.environ.get("HERMES_SERVE_HEADLESS") == "1"
-    if _headless or not WEB_DIST.exists():
-        _msg = (
-            "Headless backend (hermes serve): web UI disabled — use "
-            "`hermes dashboard` for the browser UI."
-            if _headless
-            else "Frontend not built. Run: cd web && npm run build"
-        )
-
-        @application.get("/{full_path:path}")
-        async def no_frontend(full_path: str):
-            return JSONResponse({"error": _msg}, status_code=404)
-        return
-
-    _index_path = WEB_DIST / "index.html"
-
-    def _serve_index(prefix: str = ""):
-        """Return index.html with the session token + base-path injected.
-
-        ``prefix`` is the normalised ``X-Forwarded-Prefix`` (e.g. ``/hermes``)
-        or empty string when served at root.
-
-        When the OAuth auth gate is active (``app.state.auth_required``),
-        the legacy ``_SESSION_TOKEN`` is NOT injected — the SPA reads
-        identity from ``/api/auth/me`` over cookie auth instead.  The
-        ``__HERMES_AUTH_REQUIRED__`` flag lets the SPA pick the right
-        auth scheme for /api/pty and /api/ws (ticket vs token).
-        """
-        try:
-            html = _index_path.read_text(encoding="utf-8")
-        except OSError:
-            # The dist dir existed at mount time but index.html is missing or
-            # unreadable now (partial build, wiped dist, permissions). Without
-            # this guard every request raises FileNotFoundError (500). Return
-            # the same JSON 404 payload mount_spa uses for a fully-missing
-            # dist so clients get a clear, consistent signal.
-            return JSONResponse(
-                {"error": "Frontend not built. Run: cd web && npm run build"},
-                status_code=404,
-            )
-        chat_js = "true" if _DASHBOARD_EMBEDDED_CHAT_ENABLED else "false"
-        gated = bool(getattr(app.state, "auth_required", False))
-        gated_js = "true" if gated else "false"
-        if gated:
-            bootstrap_script = (
-                f"<script>"
-                f"window.__HERMES_DASHBOARD_EMBEDDED_CHAT__={chat_js};"
-                f'window.__HERMES_BASE_PATH__="{prefix}";'
-                f"window.__HERMES_AUTH_REQUIRED__={gated_js};"
-                f"</script>"
-            )
-        else:
-            bootstrap_script = (
-                f'<script>window.__HERMES_SESSION_TOKEN__="{_SESSION_TOKEN}";'
-                f"window.__HERMES_DASHBOARD_EMBEDDED_CHAT__={chat_js};"
-                f'window.__HERMES_BASE_PATH__="{prefix}";'
-                f"window.__HERMES_AUTH_REQUIRED__={gated_js};"
-                f"</script>"
-            )
-        if prefix:
-            # Rewrite absolute asset URLs baked into the Vite build so the
-            # browser fetches them through the same proxy prefix.
-            html = html.replace('href="/assets/', f'href="{prefix}/assets/')
-            html = html.replace('src="/assets/', f'src="{prefix}/assets/')
-            html = html.replace('href="/favicon.ico"', f'href="{prefix}/favicon.ico"')
-            html = html.replace('href="/fonts/', f'href="{prefix}/fonts/')
-            html = html.replace('href="/ds-assets/', f'href="{prefix}/ds-assets/')
-            html = html.replace('src="/ds-assets/', f'src="{prefix}/ds-assets/')
-        # Theme flash mitigation: when the active theme is a user theme
-        # (``HERMES_HOME/dashboard-themes/<name>.yaml``), inject a minimal
-        # critical-CSS block so the first paint uses the target palette.
-        # Without this the SPA paints the default Hermes Teal canvas, then
-        # ``ThemeProvider`` flips the CSS variables once
-        # ``/api/dashboard/themes`` resolves.  Built-in themes are already
-        # in the bundle's ``presets.ts`` so no shim is needed for them.
-        theme_bootstrap = _render_active_theme_bootstrap_css()
-        if theme_bootstrap:
-            html = html.replace("</head>", f"{theme_bootstrap}</head>", 1)
-        html = html.replace("</head>", f"{bootstrap_script}</head>", 1)
-        return HTMLResponse(
-            html,
-            headers={"Cache-Control": "no-store, no-cache, must-revalidate"},
-        )
-
-    # When served behind a path-prefix proxy, the built CSS contains
-    # absolute ``url(/fonts/...)`` and ``url(/ds-assets/...)`` references.
-    # Browsers resolve those against the document origin, which means
-    # under ``/hermes`` they'd hit ``mission-control.tilos.com/fonts/...``
-    # (the MC Pages app), not the Hermes backend. Intercept CSS asset
-    # requests BEFORE the StaticFiles mount and rewrite the absolute paths
-    # when a prefix is in play.
-    @application.get("/assets/{filename}.css")
-    async def serve_css(filename: str, request: Request):
-        css_path = WEB_DIST / "assets" / f"{filename}.css"
-        if not css_path.is_file() or not css_path.resolve().is_relative_to(
-            WEB_DIST.resolve()
-        ):
-            return JSONResponse({"error": "not found"}, status_code=404)
-        prefix = _normalise_prefix(request.headers.get("x-forwarded-prefix"))
-        css = css_path.read_text(encoding="utf-8")
-        if prefix:
-            for asset_dir in ("/fonts/", "/fonts-terminal/", "/ds-assets/", "/assets/"):
-                css = css.replace(f"url({asset_dir}", f"url({prefix}{asset_dir}")
-                css = css.replace(f"url(\"{asset_dir}", f"url(\"{prefix}{asset_dir}")
-                css = css.replace(f"url('{asset_dir}", f"url('{prefix}{asset_dir}")
-        return Response(
-            content=css,
-            media_type="text/css",
-            headers={"Cache-Control": _IMMUTABLE_ASSET_CACHE_CONTROL},
-        )
-
-    class _ImmutableAssetFiles(StaticFiles):
-        """StaticFiles that marks hashed bundle assets immutable.
-
-        Everything under ``/assets/`` carries a Vite content hash in its
-        filename, so a given URL's bytes can never change — a rebuild
-        produces a NEW filename referenced by a fresh (``no-store``)
-        index.html. Without this header every dashboard load re-validated
-        each chunk; with it the browser serves reloads straight from its
-        HTTP cache.
-        """
-
-        async def get_response(self, path: str, scope):
-            response = await super().get_response(path, scope)
-            if response.status_code == 200:
-                response.headers["Cache-Control"] = _IMMUTABLE_ASSET_CACHE_CONTROL
-            return response
-
-    application.mount(
-        "/assets", _ImmutableAssetFiles(directory=WEB_DIST / "assets"), name="assets"
-    )
-
-    @application.get("/{full_path:path}")
-    async def serve_spa(full_path: str, request: Request):
-        prefix = _normalise_prefix(request.headers.get("x-forwarded-prefix"))
-        # An unmatched /api/* path is a missing/renamed endpoint, NOT a
-        # client-side route. Falling through to index.html here returns
-        # `<!doctype html>` with status 200, which makes JSON clients (the
-        # desktop app's fetchJson, dashboard fetch wrappers) blow up with an
-        # opaque `SyntaxError: Unexpected token '<'`. Return a real 404 JSON
-        # so the caller sees a clear "no such endpoint" instead.
-        if full_path == "api" or full_path.startswith("api/"):
-            return JSONResponse(
-                {"detail": f"No such API endpoint: /{full_path}"},
-                status_code=404,
-            )
-        file_path = WEB_DIST / full_path
-        # Prevent path traversal via url-encoded sequences (%2e%2e/)
-        if (
-            full_path
-            and file_path.resolve().is_relative_to(WEB_DIST.resolve())
-            and file_path.exists()
-            and file_path.is_file()
-        ):
-            return FileResponse(file_path)
-        return _serve_index(prefix)
 
 
 # ---------------------------------------------------------------------------
@@ -16814,6 +15779,10 @@ _mount_plugin_api_routes()
 # not whether the routes exist.
 from hermes_cli.dashboard_auth.routes import router as _dashboard_auth_router  # noqa: E402
 app.include_router(_dashboard_auth_router)
+
+# /api/console moved to console_ws_mixin; register the route here so it
+# stays ahead of the SPA catch-all mounted below.
+app.websocket("/api/console")(console_ws)
 
 mount_spa(app)
